Add new CodeInput
