+++ desktop/cmp/input/CodeInput.scss
@@ -10,7 +10,7 @@
   -webkit-box-shadow: inset 0 0 0 0 rgba(16, 22, 26, 0.4);
 }
 
-.xh-json-input {
+.xh-code-input {
   flex-direction: column;
 
   &.xh-input-invalid {
@@ -58,7 +58,7 @@
   }
 }
 
-.xh-form-field-inline .xh-json-input {
+.xh-form-field-inline .xh-code-input {
   flex-direction: row;
 
   & > .CodeMirror {
